FIX slow form creation in display mode
This is due to calculation of possible values from rdfs:range of each property.

Until recently, it was OK, but I recently added the calculation of a label for URI in triples.
And then in unit test one obtains (TestTableView) 45 seconds for the display of my FOAF.

There are at least two problems:
- 1. the calculation of the possible values is also called in display only, where it is useless;
- 2. for each triple, the possible values are re-calculated, although they depend on the property only.

This fixes problem 1.